fix an error recognising a client process in a trace event and change int to long because the counters exceeded Integer.MAX

+++ src/org/executequery/gui/browser/managment/tracemanager/TableCounterModel.java
@@ -87,7 +87,7 @@ public class TableCounterModel extends AbstractTableModel {
         if (visibleColumnNameFromIndex(columnIndex).contentEquals(LogConstants.TABLE))
             return String.class;
         else
-            return Integer.class;
+            return Long.class;
     }
 
     private String visibleColumnNameFromIndex(int i) {
